docker/django/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.3
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
redis==6.2.0
requests==2.32.4
spotipy==2.25.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0

kafka-python==2.2.11
